+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - EntityPanel.DetailLayout.layout(), entityPanel parameter removed, now returns Optional, related changes.
 - EntityPanel.DetailController reintroduced, related changes.
 - WindowDetailLayout.WindowDetailController.select() bug fixed, null check added.
+- WindowDetailLayout static factory methods removed.
 
 ## 0.17.35
 ### is.codion.common.model